+++ Android/Campong/app/src/main/AndroidManifest.xml
@@ -13,7 +13,9 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.Campong"
         android:usesCleartextTraffic="true">
-        <activity android:name=".view.SaveMyPostActivity" />
+        <activity
+            android:name=".view.WritePostActivity"
+            android:windowSoftInputMode="adjustResize" />
         <activity android:name=".view.SendRequestActivity" />
         <activity android:name=".view.EditProfileActivity" />
         <activity android:name=".view.SearchActivity" />
