--- Android/Campong/app/src/main/java/com/basecamp/campong/view/SaveMyPostActivity.kt
@@ -1,84 +1,0 @@
-package com.basecamp.campong.view
-
-import android.app.Activity
-import android.content.Intent
-import android.content.pm.PackageManager
-import android.net.Uri
-import android.os.Bundle
-import android.widget.Toast
-import androidx.appcompat.app.AlertDialog
-import androidx.appcompat.app.AppCompatActivity
-import androidx.core.content.ContextCompat
-import com.basecamp.campong.databinding.ActivitySaveMyPostBinding
-
-class SaveMyPostActivity : AppCompatActivity() {
-
-    private lateinit var mBinding: ActivitySaveMyPostBinding
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        mBinding = ActivitySaveMyPostBinding.inflate(layoutInflater)
-
-        initAddPhotoButton()
-
-        setContentView(mBinding.root)
-    }
-
-    fun initAddPhotoButton() {
-        mBinding.button3.setOnClickListener {
-            when {
-                ContextCompat.checkSelfPermission(
-                    this, android.Manifest.permission.READ_EXTERNAL_STORAGE
-                ) == PackageManager.PERMISSION_GRANTED -> {
-                    navigatePhotos()
-                }
-                shouldShowRequestPermissionRationale(android.Manifest.permission.READ_EXTERNAL_STORAGE) -> {
-                    showPermissionContextPopup()
-                }
-
-                else -> {
-                    requestPermissions(
-                        arrayOf(android.Manifest.permission.READ_EXTERNAL_STORAGE), 1000
-                    )
-                }
-            }
-        }
-    }
-
-    private fun showPermissionContextPopup() {
-        AlertDialog.Builder(this)
-            .setMessage("앱에서 사진을 가져오기 위해서는 저장소 권한이 필요합니다.")
-            .setPositiveButton("확인") { _, _ ->
-                requestPermissions(
-                    arrayOf(android.Manifest.permission.READ_EXTERNAL_STORAGE), 1000
-                )
-            }
-            .setNegativeButton("취소") { _, _ -> }
-            .create().show()
-    }
-
-    private fun navigatePhotos() {
-        val intent = Intent(Intent.ACTION_PICK)
-        intent.type = "image/*"
-        startActivityForResult(intent, 2000)
-    }
-
-    override fun onActivityResult(requestCode: Int, resultCode: Int, data: Intent?) {
-        super.onActivityResult(requestCode, resultCode, data)
-
-        if (resultCode != Activity.RESULT_OK) return
-
-        when (requestCode) {
-            2000 -> {
-                val selectedImageUri: Uri? = data?.data
-                if (selectedImageUri != null) {
-                    mBinding.button3.setImageURI(selectedImageUri)
-                } else {
-                    Toast.makeText(this, "사진을 가져오지 못했습니다.", Toast.LENGTH_SHORT).show()
-                }
-            }
-            else -> {
-                Toast.makeText(this, "사진을 가져오지 못했습니다.", Toast.LENGTH_SHORT).show()
-            }
-        }
-    }
-}
