+++ Android/Campong/app/src/main/res/values/strings.xml
@@ -1,9 +1,4 @@
 <resources>
     <string name="app_name">Campong</string>
-    <string name="title_activity_search">SearchActivity</string>
-    <!--
-    This string is used for square devices and overridden by hello_world in
-    values-round/strings.xml for round devices.
-    -->
-    <string name="hello_world">Hello Square World!</string>
+    <string name="up_button">up_button</string>
 </resources>
